requirements.txt
numpy
matplotlib
opencv-python
torch
torchvision
facenet-pytorch
pyspark
fpdf
Pillow
